fix(refactor): hierarchy
