+++ substrait_consumer/functional/comparison_configs.py
@@ -1,5 +1,5 @@
-from tests.functional.queries.sql.comparison_functions_sql import SQL_SCALAR
-from tests.functional.queries.ibis_expressions.comparison_functions_expr import (
+from substrait_consumer.functional.queries.sql.comparison_functions_sql import SQL_SCALAR
+from substrait_consumer.functional.queries.ibis_expressions.comparison_functions_expr import (
     IBIS_SCALAR,
 )
