+++ substrait_consumer/functional/queries/ibis_expressions/arithmetic_functions_expr.py
