+++ substrait_consumer/functional/queries/ibis_expressions/boolean_functions_expr.py
