+++ substrait_consumer/functional/queries/ibis_expressions/logarithmic_functions_expr.py
