+++ substrait_consumer/functional/queries/ibis_expressions/string_functions_expr.py
