+++ substrait_consumer/functional/queries/sql/arithmetic_demical_functions_sql.py
