+++ substrait_consumer/functional/queries/sql/string_functions_sql.py
