+++ substrait_consumer/functional/rounding_configs.py
@@ -1,5 +1,5 @@
-from tests.functional.queries.ibis_expressions.rounding_functions_expr import IBIS_SCALAR
-from tests.functional.queries.sql.rounding_functions_sql import SQL_SCALAR
+from substrait_consumer.functional.queries.ibis_expressions.rounding_functions_expr import IBIS_SCALAR
+from substrait_consumer.functional.queries.sql.rounding_functions_sql import SQL_SCALAR
 
 SCALAR_FUNCTIONS = (
     {
