+++ substrait_consumer/tests/functional/extension_functions/test_arithmetic_functions.py
@@ -4,16 +4,16 @@ import duckdb
 from ibis.expr.types.relations import Table
 from ibis_substrait.tests.compiler.conftest import *
 
-from tests.functional.logarithmic_tests import SCALAR_FUNCTIONS
-from tests.parametrization import custom_parametrization
-from tests.functional.common import substrait_function_test
+from substrait_consumer.functional.arithmetic_configs import AGGREGATE_FUNCTIONS, SCALAR_FUNCTIONS
+from substrait_consumer.functional.common import load_custom_duckdb_table, substrait_function_test
+from substrait_consumer.parametrization import custom_parametrization
 
 
 @pytest.mark.usefixtures("prepare_tpch_parquet_data")
-class TestLogarithmicFunctions:
+class TestArithmeticFunctions:
     """
     Test Class verifying different consumers are able to run substrait plans
-    that include substrait logarithmic functions.
+    that include substrait arithmetic functions.
     """
 
     @staticmethod
@@ -24,14 +24,20 @@ class TestLogarithmicFunctions:
         cls.db_connection = duckdb.connect()
         cls.db_connection.execute("install substrait")
         cls.db_connection.execute("load substrait")
+        load_custom_duckdb_table(cls.db_connection)
+        cls.table_t = ibis.table(
+            [("a", dt.int32), ("b", dt.int32), ("c", dt.boolean), ("d", dt.boolean)],
+            name="t",
+        )
+
         cls.created_tables = set()
 
         yield
 
         cls.db_connection.close()
 
-    @custom_parametrization(SCALAR_FUNCTIONS)
-    def test_logarithmic_functions(
+    @custom_parametrization(SCALAR_FUNCTIONS + AGGREGATE_FUNCTIONS)
+    def test_arithmetic_functions(
         self,
         test_name: str,
         file_names: Iterable[str],
@@ -40,6 +46,7 @@ class TestLogarithmicFunctions:
         producer,
         consumer,
         partsupp,
+        lineitem,
     ) -> None:
         substrait_function_test(
             self.db_connection,
@@ -49,5 +56,7 @@ class TestLogarithmicFunctions:
             ibis_expr,
             producer,
             consumer,
-            partsupp
+            partsupp,
+            lineitem,
+            self.table_t,
         )
