+++ substrait_consumer/tests/functional/extension_functions/test_boolean_functions.py
@@ -4,21 +4,21 @@ import duckdb
 from ibis.expr.types.relations import Table
 from ibis_substrait.tests.compiler.conftest import *
 
-from tests.functional.arithmetic_tests import AGGREGATE_FUNCTIONS, SCALAR_FUNCTIONS
-from tests.functional.common import load_custom_duckdb_table, substrait_function_test
-from tests.parametrization import custom_parametrization
+from substrait_consumer.functional.boolean_configs import AGGREGATE_FUNCTIONS, SCALAR_FUNCTIONS
+from substrait_consumer.functional.common import load_custom_duckdb_table, substrait_function_test
+from substrait_consumer.parametrization import custom_parametrization
 
 
 @pytest.mark.usefixtures("prepare_tpch_parquet_data")
-class TestArithmeticFunctions:
+class TestBooleanFunctions:
     """
     Test Class verifying different consumers are able to run substrait plans
-    that include substrait arithmetic functions.
+    that include substrait boolean functions.
     """
 
     @staticmethod
-    @pytest.fixture(scope="class", autouse=True)
-    def setup_teardown_class(request):
+    @pytest.fixture(scope="function", autouse=True)
+    def setup_teardown_function(request):
         cls = request.cls
 
         cls.db_connection = duckdb.connect()
@@ -37,7 +37,7 @@ class TestArithmeticFunctions:
         cls.db_connection.close()
 
     @custom_parametrization(SCALAR_FUNCTIONS + AGGREGATE_FUNCTIONS)
-    def test_arithmetic_functions(
+    def test_boolean_functions(
         self,
         test_name: str,
         file_names: Iterable[str],
@@ -45,8 +45,6 @@ class TestArithmeticFunctions:
         ibis_expr: Callable[[Table], Table],
         producer,
         consumer,
-        partsupp,
-        lineitem,
     ) -> None:
         substrait_function_test(
             self.db_connection,
@@ -56,7 +54,5 @@ class TestArithmeticFunctions:
             ibis_expr,
             producer,
             consumer,
-            partsupp,
-            lineitem,
             self.table_t,
         )
