+++ substrait_consumer/tests/functional/extension_functions/test_comparison_functions.py
@@ -4,16 +4,16 @@ import duckdb
 from ibis.expr.types.relations import Table
 from ibis_substrait.tests.compiler.conftest import *
 
-from tests.functional.datetime_tests import SCALAR_FUNCTIONS
-from tests.parametrization import custom_parametrization
-from tests.functional.common import substrait_function_test
+from substrait_consumer.functional.common import load_custom_duckdb_table, substrait_function_test
+from substrait_consumer.functional.comparison_configs import SCALAR_FUNCTIONS
+from substrait_consumer.parametrization import custom_parametrization
 
 
 @pytest.mark.usefixtures("prepare_tpch_parquet_data")
-class TestDatetimeFunctions:
+class TestComparisonFunctions:
     """
     Test Class verifying different consumers are able to run substrait plans
-    that include substrait datetime functions.
+    that include substrait comparison functions.
     """
 
     @staticmethod
@@ -24,6 +24,8 @@ class TestDatetimeFunctions:
         cls.db_connection = duckdb.connect()
         cls.db_connection.execute("install substrait")
         cls.db_connection.execute("load substrait")
+        load_custom_duckdb_table(cls.db_connection)
+
         cls.created_tables = set()
 
         yield
@@ -31,7 +33,7 @@ class TestDatetimeFunctions:
         cls.db_connection.close()
 
     @custom_parametrization(SCALAR_FUNCTIONS)
-    def test_datetime_functions(
+    def test_comparison_functions(
         self,
         test_name: str,
         file_names: Iterable[str],
@@ -40,6 +42,7 @@ class TestDatetimeFunctions:
         producer,
         consumer,
         partsupp,
+        nation,
     ) -> None:
         substrait_function_test(
             self.db_connection,
@@ -49,5 +52,6 @@ class TestDatetimeFunctions:
             ibis_expr,
             producer,
             consumer,
-            partsupp
+            partsupp,
+            nation,
         )
