+++ substrait_consumer/tests/functional/extension_functions/test_datetime_functions.py
@@ -4,16 +4,16 @@ import duckdb
 from ibis.expr.types.relations import Table
 from ibis_substrait.tests.compiler.conftest import *
 
-from tests.functional.boolean_tests import AGGREGATE_FUNCTIONS, SCALAR_FUNCTIONS
-from tests.functional.common import load_custom_duckdb_table, substrait_function_test
-from tests.parametrization import custom_parametrization
+from substrait_consumer.functional.datetime_configs import SCALAR_FUNCTIONS
+from substrait_consumer.parametrization import custom_parametrization
+from substrait_consumer.functional.common import substrait_function_test
 
 
 @pytest.mark.usefixtures("prepare_tpch_parquet_data")
-class TestBooleanFunctions:
+class TestDatetimeFunctions:
     """
     Test Class verifying different consumers are able to run substrait plans
-    that include substrait boolean functions.
+    that include substrait datetime functions.
     """
 
     @staticmethod
@@ -24,20 +24,14 @@ class TestBooleanFunctions:
         cls.db_connection = duckdb.connect()
         cls.db_connection.execute("install substrait")
         cls.db_connection.execute("load substrait")
-        load_custom_duckdb_table(cls.db_connection)
-        cls.table_t = ibis.table(
-            [("a", dt.int32), ("b", dt.int32), ("c", dt.boolean), ("d", dt.boolean)],
-            name="t",
-        )
-
         cls.created_tables = set()
 
         yield
 
         cls.db_connection.close()
 
-    @custom_parametrization(SCALAR_FUNCTIONS + AGGREGATE_FUNCTIONS)
-    def test_boolean_functions(
+    @custom_parametrization(SCALAR_FUNCTIONS)
+    def test_datetime_functions(
         self,
         test_name: str,
         file_names: Iterable[str],
@@ -45,6 +39,7 @@ class TestBooleanFunctions:
         ibis_expr: Callable[[Table], Table],
         producer,
         consumer,
+        partsupp,
     ) -> None:
         substrait_function_test(
             self.db_connection,
@@ -54,5 +49,5 @@ class TestBooleanFunctions:
             ibis_expr,
             producer,
             consumer,
-            self.table_t,
+            partsupp
         )
