+++ substrait_consumer/tests/functional/extension_functions/test_substrait_function_names.py
@@ -5,12 +5,11 @@ import duckdb
 from ibis.expr.types.relations import Table
 from ibis_substrait.tests.compiler.conftest import *
 
-from tests.consumers import DuckDBConsumer
-from tests.functional import (
-    arithmetic_tests, boolean_tests, comparison_tests, datetime_tests, logarithmic_tests,
-    rounding_tests)
-from tests.functional.common import check_subtrait_function_names, load_custom_duckdb_table
-from tests.parametrization import custom_parametrization
+from substrait_consumer.consumers import DuckDBConsumer
+from substrait_consumer.functional import (
+    arithmetic_configs, boolean_configs, comparison_configs, datetime_configs, logarithmic_configs, rounding_configs)
+from substrait_consumer.functional.common import check_subtrait_function_names, load_custom_duckdb_table
+from substrait_consumer.parametrization import custom_parametrization
 
 
 @pytest.mark.usefixtures("prepare_tpch_parquet_data")
@@ -40,7 +39,7 @@ class TestSubstraitFunctionNames:
         cls.db_connection.close()
 
     @custom_parametrization(
-        arithmetic_tests.SCALAR_FUNCTIONS + arithmetic_tests.AGGREGATE_FUNCTIONS
+        arithmetic_configs.SCALAR_FUNCTIONS + arithmetic_configs.AGGREGATE_FUNCTIONS
     )
     def test_arithmetic_function_names(
         self,
@@ -67,7 +66,7 @@ class TestSubstraitFunctionNames:
         )
 
     @custom_parametrization(
-        boolean_tests.SCALAR_FUNCTIONS + boolean_tests.AGGREGATE_FUNCTIONS
+        boolean_configs.SCALAR_FUNCTIONS + boolean_configs.AGGREGATE_FUNCTIONS
     )
     def test_boolean_function_names(
         self,
@@ -84,7 +83,7 @@ class TestSubstraitFunctionNames:
             test_name, file_names, sql_query, ibis_expr, producer, self.table_t
         )
 
-    @custom_parametrization(comparison_tests.SCALAR_FUNCTIONS)
+    @custom_parametrization(comparison_configs.SCALAR_FUNCTIONS)
     def test_comparison_function_names(
         self,
         test_name: str,
@@ -102,7 +101,7 @@ class TestSubstraitFunctionNames:
             test_name, file_names, sql_query, ibis_expr, producer, partsupp, nation
         )
 
-    @custom_parametrization(datetime_tests.SCALAR_FUNCTIONS)
+    @custom_parametrization(datetime_configs.SCALAR_FUNCTIONS)
     def test_datetime_function_names(
         self,
         test_name: str,
@@ -119,7 +118,7 @@ class TestSubstraitFunctionNames:
             test_name, file_names, sql_query, ibis_expr, producer, partsupp
         )
 
-    @custom_parametrization(logarithmic_tests.SCALAR_FUNCTIONS)
+    @custom_parametrization(logarithmic_configs.SCALAR_FUNCTIONS)
     def test_logarithmic_function_names(
         self,
         test_name: str,
@@ -136,7 +135,7 @@ class TestSubstraitFunctionNames:
             test_name, file_names, sql_query, ibis_expr, producer, partsupp
         )
 
-    @custom_parametrization(rounding_tests.SCALAR_FUNCTIONS)
+    @custom_parametrization(rounding_configs.SCALAR_FUNCTIONS)
     def test_rounding_function_names(
         self,
         test_name: str,
